refactoring: replaced type User into user segment of shared layer.
