+++ src/features/users/model/types.ts
@@ -1,23 +1,9 @@
 import { Styles } from '@/shared/mui';
+import { Photo, User } from '@/shared/user';
 
 export interface UsersQueryParams {
   usersQueryCount: number;
   currentPage: number;
-}
-
-export type Photo = string | null;
-
-export interface Photos {
-  small: Photo;
-  large: Photo;
-}
-
-export interface User {
-  id: number | string;
-  name: string;
-  status: string;
-  photos: Photos;
-  followed: boolean;
 }
 
 export interface UsersResponse {
